aura_agent/requirements.txt
pvporcupine
speechrecognition
pyaudio
google-generativeai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow